refactor: module layout
cleaner structure and allows for more non-underscore-prefixed module names.
