+++ src/unxt/__init__.py
@@ -3,30 +3,17 @@
 Copyright (c) 2023 Galactic Dynamics. All rights reserved.
 """
 
-from . import (
-    _quantity,
+from . import unitsystems
+from ._unxt import (
     experimental,  # noqa: F401
-    unitsystems,
+    quantity,
 )
-from ._optional_deps import HAS_ASTROPY, HAS_GALA, HAS_MATPLOTLIB
-from ._quantity import *
+from ._unxt.quantity import *
 from ._version import version as __version__
 from .unitsystems import AbstractUnitSystem, UnitSystem, unitsystem
 
-# Register interoperability
 # isort: split
-if HAS_ASTROPY:
-    from ._interop import unxt_interop_astropy
-
-    del unxt_interop_astropy
-if HAS_GALA:
-    from ._interop import unxt_interop_gala
-
-    del unxt_interop_gala
-if HAS_MATPLOTLIB:
-    from ._interop import unxt_interop_mpl as interop_mpl
-
-    interop_mpl.setup_matplotlib_support_for_unxt(enable=True)
+from . import _interop  # noqa: F401  # register interop
 
 __all__ = [
     "__version__",
@@ -36,7 +23,7 @@ __all__ = [
     "UnitSystem",  # main user-facing class
     "unitsystem",  # convenience constructor
 ]
-__all__ += _quantity.__all__
+__all__ += quantity.__all__
 
 # Clean up namespace
-del HAS_ASTROPY, HAS_GALA, HAS_MATPLOTLIB, _quantity
+del quantity
