+++ src/unxt/_unxt/__init__.py
@@ -1,0 +1,6 @@
+"""unxt: Quantities in JAX.
+
+Copyright (c) 2023 Galactic Dynamics. All rights reserved.
+"""
+
+__all__: list[str] = []
