+++ src/unxt/_unxt/quantity/base.py
@@ -24,7 +24,7 @@ import quaxed.numpy as jnp
 import quaxed.operator as qoperator
 from quaxed.array_api._dispatch import dispatcher
 
-from unxt._units import Unit
+from unxt._unxt.units import Unit
 
 if TYPE_CHECKING:
     from array_api import ArrayAPINamespace
