+++ src/unxt/_unxt/quantity/fast.py
