+++ src/unxt/_unxt/unitsystems/__init__.py
@@ -1,0 +1,3 @@
+"""Tools for representing systems of units using ``astropy.units``."""
+
+__all__: list[str] = []
