+++ src/unxt/_unxt/unitsystems/utils.py
@@ -6,11 +6,11 @@ __all__ = ["unitsystem"]
 import astropy.units as u
 from plum import dispatch
 
-from ._base import AbstractUnitSystem
-from ._core import DimensionlessUnitSystem, UnitSystem
-from ._realizations import NAMED_UNIT_SYSTEMS, dimensionless
-from unxt._quantity.base import AbstractQuantity
-from unxt._typing import Unit
+from .base import AbstractUnitSystem
+from .core import DimensionlessUnitSystem, UnitSystem
+from .realizations import NAMED_UNIT_SYSTEMS, dimensionless
+from unxt._unxt.quantity.base import AbstractQuantity
+from unxt._unxt.typing_ext import Unit
 
 
 @dispatch
