--- src/unxt/unitsystems/__init__.py
@@ -1,16 +1,0 @@
-"""Tools for representing systems of units using ``astropy.units``."""
-
-from . import _base, _core, _realizations, _utils
-from ._base import *
-from ._core import *
-from ._realizations import *
-from ._utils import *
-
-__all__: list[str] = []
-__all__ += _base.__all__
-__all__ += _core.__all__
-__all__ += _realizations.__all__
-__all__ += _utils.__all__
-
-# Clean up namespace
-del _core, _utils, _realizations
